executor: split ddl test into new package
