+++ executor/BUILD.bazel
@@ -297,7 +297,6 @@ go_test(
         "concurrent_map_test.go",
         "copr_cache_test.go",
         "cte_test.go",
-        "ddl_test.go",
         "delete_test.go",
         "distsql_test.go",
         "executor_failpoint_test.go",
@@ -372,8 +371,6 @@ go_test(
         "//config",
         "//ddl",
         "//ddl/placement",
-        "//ddl/schematracker",
-        "//ddl/testutil",
         "//ddl/util",
         "//distsql",
         "//domain",
@@ -387,7 +384,6 @@ go_test(
         "//expression/aggregation",
         "//infoschema",
         "//kv",
-        "//meta",
         "//meta/autoid",
         "//metrics",
         "//parser",
@@ -406,7 +402,6 @@ go_test(
         "//sessionctx/binloginfo",
         "//sessionctx/stmtctx",
         "//sessionctx/variable",
-        "//sessionctx/variable/featuretag/disttask",
         "//sessiontxn",
         "//sessiontxn/staleread",
         "//statistics",
@@ -424,7 +419,6 @@ go_test(
         "//testkit/testdata",
         "//testkit/testmain",
         "//testkit/testsetup",
-        "//testkit/testutil",
         "//types",
         "//util",
         "//util/benchdaily",
