+++ executor/test/ddl/BUILD.bazel
@@ -1,0 +1,38 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "ddl_test",
+    timeout = "short",
+    srcs = ["ddl_test.go"],
+    flaky = True,
+    shard_count = 42,
+    deps = [
+        "//ddl/schematracker",
+        "//ddl/testutil",
+        "//ddl/util",
+        "//domain",
+        "//errno",
+        "//infoschema",
+        "//kv",
+        "//meta",
+        "//meta/autoid",
+        "//parser/model",
+        "//parser/mysql",
+        "//parser/terror",
+        "//planner/core",
+        "//sessionctx/variable",
+        "//sessionctx/variable/featuretag/disttask",
+        "//sessiontxn",
+        "//store/mockstore",
+        "//table",
+        "//table/tables",
+        "//testkit",
+        "//testkit/testutil",
+        "//types",
+        "//util/chunk",
+        "//util/dbterror",
+        "//util/dbterror/exeerrors",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+    ],
+)
